+++ src/main/java/ir/comprehensive/service/ProductDeliveryService.java
@@ -2,15 +2,12 @@ package ir.comprehensive.service;
 
 import ir.comprehensive.domain.ProductDelivery;
 import ir.comprehensive.domain.ProductStatus;
-import ir.comprehensive.mapper.ProductDeliveryMapper;
 import ir.comprehensive.model.ProductDeliveryModel;
 import ir.comprehensive.repository.ProductDeliveryRepository;
-import ir.comprehensive.service.response.RequestCallback;
-import ir.comprehensive.service.response.ResponseStatus;
+import ir.comprehensive.service.extra.GeneralException;
+import ir.comprehensive.service.extra.Swappable;
 import ir.comprehensive.utils.MessageUtils;
 import ir.comprehensive.utils.StringUtils;
-import javafx.collections.FXCollections;
-import javafx.collections.ObservableList;
 import org.springframework.data.domain.Example;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
@@ -19,29 +16,33 @@ import org.springframework.transaction.annotation.Transactional;
 import javax.persistence.criteria.Predicate;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.stream.Collectors;
+import java.util.Optional;
 
 @Service
 @Transactional
-public class ProductDeliveryService {
+public class ProductDeliveryService implements Swappable<ProductDelivery> {
     private ProductDeliveryRepository repository;
-    private ProductDeliveryMapper mapper;
 
 
-    public ProductDeliveryService(ProductDeliveryRepository repository, ProductDeliveryMapper mapper) {
+    public ProductDeliveryService(ProductDeliveryRepository repository) {
         this.repository = repository;
-        this.mapper = mapper;
     }
 
-    public ObservableList<ProductDeliveryModel> getAllModel() {
+    public Optional<ProductDelivery> load(Long id) throws GeneralException {
+        if (id == null) {
+            // TODO fix message
+            throw new GeneralException("null id");
+        }
+        return repository.findById(id);
+    }
+
+    public Optional<List<ProductDelivery>> loadByStatus(ProductStatus status) {
         ProductDelivery example = new ProductDelivery();
-        example.setStatus(ProductStatus.UNKNOWN);
-        List<ProductDelivery> all = repository.findAll(Example.of(example));
-        List<ProductDeliveryModel> collect = all.stream().map(mapper::entityToModel).collect(Collectors.toList());
-        return FXCollections.observableList(collect);
+        example.setStatus(status);
+        return Optional.of(repository.findAll(Example.of(example)));
     }
 
-    public void search(ProductDeliveryModel searchExample, RequestCallback<ObservableList<ProductDeliveryModel>> callback) {
+    public Optional<List<ProductDelivery>> search(ProductDeliveryModel searchExample) {
         Specification<ProductDelivery> productDeliverySpecification = (root, query, criteriaBuilder) -> {
             List<Predicate> predicateList = new ArrayList<>();
             if (searchExample.getPerson() != null && searchExample.getPerson().getId() != null) {
@@ -73,56 +74,51 @@ public class ProductDeliveryService {
             return criteriaBuilder.and(predicateList.toArray(new Predicate[predicateList.size()]));
         };
 
-        List<ProductDeliveryModel> allModel = repository.findAll(productDeliverySpecification).stream().map(mapper::entityToModel).collect(Collectors.toList());
-        callback.accept(FXCollections.observableList(allModel), MessageUtils.Message.SUCCESS_LOAD, ResponseStatus.SUCCESS);
+        return Optional.of(repository.findAll(productDeliverySpecification));
     }
 
-    public void saveOrUpdate(ProductDeliveryModel model, RequestCallback<ProductDelivery> callback) {
-        // Convert to entity
-        ProductDelivery productDelivery = mapper.modelToEntity(model);
-
-        // null point check after convert
+    private void validateEntity(ProductDelivery productDelivery) throws GeneralException {
         if (productDelivery == null) {
-            callback.accept(null, MessageUtils.Message.ERROR_IN_SAVE, ResponseStatus.FAIL);
-            return;
-        }
-
-        // apply save
-        if (null == productDelivery.getId()) {
-            String callbackMessage = MessageUtils.Message.PRODUCT + " " + MessageUtils.Message.SUCCESS_SAVE;
-
-            productDelivery.setStatus(ProductStatus.UNKNOWN);
-            productDelivery.setReceivedDate(null);
-
-            callback.accept(repository.save(productDelivery), callbackMessage, ResponseStatus.SUCCESS);
-            return;
-        }
-
-        // apply update
-
-        ProductDelivery loadedProductDelivery = repository.findById(productDelivery.getId()).orElse(null);
-        if (loadedProductDelivery == null) {
-            callback.accept(null, MessageUtils.Message.ERROR_IN_SAVE, ResponseStatus.FAIL);
-        } else {
-            loadedProductDelivery.setId(productDelivery.getId());
-            loadedProductDelivery.setPerson(productDelivery.getPerson());
-            loadedProductDelivery.setProduct(productDelivery.getProduct());
-            loadedProductDelivery.setDescription(productDelivery.getDescription());
-            loadedProductDelivery.setDeliveryDate(productDelivery.getDeliveryDate());
-            loadedProductDelivery.setDesiredDate(productDelivery.getDesiredDate());
-            loadedProductDelivery.setStatus(productDelivery.getStatus());
-            loadedProductDelivery.setReceivedDate(productDelivery.getStatus().equals(ProductStatus.RECEIVED) ? productDelivery.getReceivedDate() : null);
-            String callbackMessage = MessageUtils.Message.PRODUCT + " " + MessageUtils.Message.SUCCESS_UPDATE;
-            callback.accept(repository.save(loadedProductDelivery), callbackMessage, ResponseStatus.SUCCESS);
-
+            throw new GeneralException(MessageUtils.Message.ERROR_IN_SAVE);
         }
     }
 
-    public void delete(Long id) {
+    public Optional<ProductDelivery> save(ProductDelivery productDelivery) throws GeneralException {
+        validateEntity(productDelivery);
+        productDelivery.setId(null);
+        productDelivery.setStatus(ProductStatus.UNKNOWN);
+        productDelivery.setReceivedDate(null);
+        return Optional.of(repository.save(productDelivery));
+    }
+
+    public Optional<ProductDelivery> update(ProductDelivery productDelivery) throws GeneralException {
+        validateEntity(productDelivery);
+        if (productDelivery.getId() == null) {
+            // TODO must fix message
+            throw new GeneralException("not null id");
+        }
+
+        productDelivery.setReceivedDate(productDelivery.getStatus().equals(ProductStatus.RECEIVED) ? productDelivery.getReceivedDate() : null);
+
+        // TODO must fix message
+        ProductDelivery loadedProductDelivery = repository.findById(productDelivery.getId()).orElseThrow(() -> new GeneralException("not found"));
+
+        return Optional.of(repository.save(swap(productDelivery, loadedProductDelivery)));
+
+    }
+
+    public Optional<ProductDelivery> saveOrUpdate(ProductDelivery productDelivery) throws GeneralException {
+        return productDelivery.getId() == null ? save(productDelivery) : update(productDelivery);
+    }
+
+
+    public Optional<Long> delete(Long id) throws GeneralException {
+        if (id == null) {
+            // TODO fix message
+            throw new GeneralException("not null id");
+        }
+
         repository.deleteById(id);
-    }
-
-    public ProductDeliveryModel load(Long id) {
-        return repository.findById(id).map(mapper::entityToModel).orElse(null);
+        return Optional.of(id);
     }
 }
